Roll WebRTC, Libjingle to revision 7728.
R=kjellander@chromium.org
TBR=kjellander@chromium.org

Review URL: https://codereview.chromium.org/746123002

Cr-Commit-Position: refs/heads/master@{#305219}

Former-commit-id: 064170774dbb2c9544a48f6f4ea8caab2ba43594

+++ content/renderer/media/rtc_video_encoder.cc
@@ -586,7 +586,7 @@ RTCVideoEncoder::~RTCVideoEncoder() {
 
 int32_t RTCVideoEncoder::InitEncode(const webrtc::VideoCodec* codec_settings,
                                     int32_t number_of_cores,
-                                    uint32_t max_payload_size) {
+                                    size_t max_payload_size) {
   DVLOG(1) << "InitEncode(): codecType=" << codec_settings->codecType
            << ", width=" << codec_settings->width
            << ", height=" << codec_settings->height
