+++ content/renderer/media/rtc_video_encoder.h
@@ -54,7 +54,7 @@ class CONTENT_EXPORT RTCVideoEncoder
   // appropriate VEA methods.
   int32_t InitEncode(const webrtc::VideoCodec* codec_settings,
                      int32_t number_of_cores,
-                     uint32_t max_payload_size) override;
+                     size_t max_payload_size) override;
   int32_t Encode(
       const webrtc::I420VideoFrame& input_image,
       const webrtc::CodecSpecificInfo* codec_specific_info,
